move bsltest to bsl
