+++ clang-tools-extra/clang-tidy/CMakeLists.txt
@@ -42,7 +42,6 @@ endif()
 add_subdirectory(android)
 add_subdirectory(abseil)
 add_subdirectory(bsl)
-add_subdirectory(bsltest)
 add_subdirectory(boost)
 add_subdirectory(bugprone)
 add_subdirectory(cert)
