+++ clang-tools-extra/clang-tidy/bsl/CMakeLists.txt
@@ -12,6 +12,7 @@ add_clang_library(clangTidyBslModule
   EnumExplicitCheck.cpp
   EnumInitCheck.cpp
   EnumScopedCheck.cpp
+  ForLoopCheck.cpp
   FriendDeclCheck.cpp
   FunctionNameUseCheck.cpp
   LambdaImplicitCaptureCheck.cpp
