--- clang-tools-extra/clang-tidy/bsltest/CMakeLists.txt
@@ -1,14 +1,0 @@
-set(LLVM_LINK_COMPONENTS support)
-
-add_clang_library(clangTidyBslTestModule
-  BslTestTidyModule.cpp
-  ForLoopCheck.cpp
-
-  LINK_LIBS
-  clangAST
-  clangASTMatchers
-  clangBasic
-  clangLex
-  clangTidy
-  clangTidyUtils
-  )
