+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -383,6 +383,9 @@ Changes in existing checks
 Renamed checks
 ^^^^^^^^^^^^^^
 
+- The 'bsltest-for-loop' check was renamed to :doc:`bsl-for-loop
+  <clang-tidy/checks/bsl-for-loop>`
+
 - The 'fuchsia-restrict-system-headers' check was renamed to :doc:`portability-restrict-system-includes
   <clang-tidy/checks/portability-restrict-system-includes>`
